requirements.txt
pyTelegramBotAPI>=4.20.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
psycopg2-binary>=2.9.0
